requirements.txt
phonenumbers
requests
mysql-connector-python
